- changed sequence of requests after character login:   - WAS: Asset Request, Item Request, Desktop Option Request   - CHANGED TO: Item Request, Asset request, Desktop Option Request   This will enable better handling of port information in the response packets
- FCDialog
  - Added recognition of default OK and CANCEL buttons being pressed. The default handlers are now invoked and call subclass handlers if they have been registered. As per intended design.
  
- ServiceLogicBase
  - fixed a bug where the db monitoring thread was not being torn down at end of execution

- FCDatabase
  - fixed a destruction bug where the worker threads would deadlock occassionally
  
- SoftwareMgrWindow
  - Added more functionality.. need to finish this.

+++ trunk/Clients/common/FCServerObj.cpp
@@ -122,7 +122,7 @@ void FCServerObj::SendCharacterItemsRequest(size_t character_id)
 	PEPacket pkt;
 	__FCPKT_CHARACTER_ITEMS_REQUEST d;
 
-	d.character_id = character_id;
+	d.character_id = (unsigned int)character_id;
   PEPacketHelper::CreatePacket(pkt, FCPKT_COMMAND, FCMSG_CHARACTER_ITEMS_REQUEST, ST_World);
   PEPacketHelper::SetPacketData(pkt, (void*)&d, sizeof(d));
